NetworkTasks: Don’t treat missing refresh token as fatal error
This change makes it so that a refresh token is no longer required in the
data returned from a network request. Some authentication services do not
reply with a refresh token, making all requests fail. The change is to
only assign the current locker’s refresh token if it exists in the data.