Fix network errors treated as “Internal error”
This patch fixes a bug that would cause all network related errors (such
as when the user is offline) as opaque `OhMyAuthError.internalError`, which
prevents the appropriate error handling to be done in clients.

The problem was that the logic dealing with responses was first checking
if a valid `HTTPURLResponse` was received before checking the error. This
patch simply flips the order of those two operations.

Tests have also been implemented for `NetworkRequestable`.